déplacement de la sélection par défaut

+++ ScreenReader.sh
@@ -14,17 +14,12 @@ MOUSE=false
 CURRENT_WINDOW=false
 SELECTION=false
 
-# Initialisation curseur
-eval $(xdotool getmouselocation --shell)
-Xo=$X
-Yo=$Y
-
 #Options choice
 case $1 in
     souris)
         MOUSE=true;;
 
-    fenCourante)
+    fen)
         CURRENT_WINDOW=true;;
 
     selection)
@@ -39,42 +34,40 @@ esac
 
 #Make screenshots
 if $MOUSE; then
-    
-    # Récupération theme actuel
-    CONFIG=$HOME/.config/kcminputrc
-    CURRENT=`kreadconfig5 --file $CONFIG --group Mouse --key cursorTheme`
-    # changement de theme
-    kwriteconfig5 --file $CONFIG --group Mouse --key cursorTheme MonThemeCurseur
-    # Rechargement du thème plasma
-    #kquitapp5 plasmashell
-    #kstart5 plasmashell
-    
+    # for distinguishing " ", "\t" from "\n"
+    IFS=''
     #Attente retour utilisateur
-    google_speech -l fr-fr  "Appuyer sur n'importe qu'elle touche pour continuer..."
-    read -p "Appuyer sur n'importe qu'elle touche pour continuer... " -n1 -s
-    #Rectangle notation :
-    # (0,1); (1,1)
-    # (0,0); (1,0)
-
-    #Offset the screenshot rectangle to the upper-left corner (0,1) by this amount in the X/Y direction
-    #Might be a problem for i18n & RTL languages but... Oh, well
-    NegX01Offset=128
-    NegY01Offset=50
-    #Size of the rectangle
-    posX10=256
-    posY10=100
-    #Set X,Y to the Cursor position.
-    eval $(xdotool getmouselocation --shell)
-    posX01=$(($X-$NegX01Offset))
-    posY01=$(($Y-$NegY01Offset))
-    #Drawing Rectangle
-    lemonbar -n my_lemonbar -d -g "$posX10"x"$posY10"+"$posX01"+"$posY01" -B '#88000000' &
-    #ScreenShot
-    scrot -a "$posX01","$posY01","$posX10","$posY10" -o "$workDir"/OCR.png
-    # Remise du theme de Curseur à l'origine
-    kwriteconfig5 --file $CONFIG --group Mouse --key cursorTheme $CURRENT
-    #kquitapp5 plasmashell
-    #kstart5 plasmashell
+    google_speech -l fr-fr  "Appuyer sur espace pour déplacer la sélection..."
+    google_speech -l fr-fr  "Et appuyer sur la touche entrée pour valider la sélection..."
+    #initialisation de la variable
+    pressed="a"
+    #lancement de la détection de la validation
+    while [ "$pressed" != "" ]; do
+        #détection de l'appui des touches
+        read -n1 -r -p "Appuyer sur espace pour déplacer la sélection et appuyer sur la touche entrée pour valider la sélection..." pressed
+        #To quit when validate
+        if [ "$pressed" == "" ]; then
+            break
+        fi        
+        #Offset the screenshot rectangle to the upper-left corner (0,1) by this amount in the X/Y direction
+        #Might be a problem for i18n & RTL languages but... Oh, well
+        NegX01Offset=128
+        NegY01Offset=50
+        #Size of the rectangle
+        posX10=256
+        posY10=100
+        #Set X,Y to the Cursor position.
+        eval $(xdotool getmouselocation --shell)
+        posX01=$(($X-$NegX01Offset))
+        posY01=$(($Y-$NegY01Offset))
+        #To take screenShot
+        scrot -a "$posX01","$posY01","$posX10","$posY10" -o "$workDir"/OCR.png
+        #Drawing Rectangle (remove old)
+        kill $(pgrep -f "lemonbar -n my_lemonbar")
+        kill $(pgrep -f "lemonbar -n my_lemonbar")
+        #Drawing Rectangle (real draw)
+        lemonbar -n my_lemonbar -d -g "$posX10"x"$posY10"+"$posX01"+"$posY01" -B '#88ff0000' &
+    done
 fi
 
 if $CURRENT_WINDOW; then
@@ -103,8 +96,10 @@ google_speech -l fr-fr "$(cat "$workDir"/text.txt)"
 
 #kill rectangle
 sleep 3
+if $MOUSE; then
 kill $(pgrep -f "lemonbar -n my_lemonbar")
 kill $(pgrep -f "lemonbar -n my_lemonbar")
+fi
 
 #delete work folder
 rm -rf "$workDir"
